fixed several compiler warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1473 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/assembly/ace/AbstractAceFileWriter.java
@@ -99,7 +99,7 @@ abstract class AbstractAceFileWriter implements AceFileWriter{
 		try{
 			while(readIterator.hasNext()){
 				AceAssembledRead read = readIterator.next();
-				int startOffset = (int)consensusSequence.getUngappedOffsetFor((int)read.getGappedStartOffset());
+				int startOffset = consensusSequence.getUngappedOffsetFor((int)read.getGappedStartOffset());
 				QualitySequence ungappedQualities;
 				try {
 					ungappedQualities = AssemblyUtil.getUngappedComplementedValidRangeQualities(read,phdDataStore.get(read.getId()).getQualitySequence());
